Refactor FormPresenter.vue for dynamic margins and extraction
Made adjustments in FormPresenter.vue to use a function that calculates dynamic margin-left spacing for form fields dependent on inherit depth to improve the layout. Also refactored the 'show_field' function for clarity and efficiency by extracting a repetitive task of searching for a field by descriptor into a new function called 'find_field_by_descriptor'. This not only makes the code cleaner but also offers easier maintenance and potential for reusability in the future.