Add logging and improve null checking, fix no right category bug
Added logging to the 'check_days' method in the TimetableManager for better tracking of the system. Enhanced the null checking in the medsenger blueprint and the AlgorithmManager, preventing potential errors during the form algorithm attachment and the categorization process. Removed versions from the requirements to allow for the latest packages. Created a script that updates all categories for algorithms.