Enhance range field configuration and functionality
Updated the condition in editors/parts/Field.vue to allow non-admin users to configure 'range' fields. Additionally, added ability for users to specify a 'step' value when field type is 'range'. In FormPresenter.vue, changed the range input step attribute to be dynamic, based on the field's set 'max' value.