Refactored FormManager to accept subcategories in fields
The changes in FormManager.py enable the acceptance of 'subcategory' when provided in the 'params' object of a field. This update will be useful when categories such as 'symptom' or 'side effect' require further specification or granularity, thus enhancing the precision of data inputs, especially in the context of heat map visualization.