Add 'fill form' feature for doctors
The main change in this commit is to add the ability for doctors to fill out forms on behalf of patients. This feature required changes across multiple files, including the FormPresenter.vue and FormCard.vue components to display the option, and updates in the managers/FormManager.py, tasks/tasks.py and blueprints/api.py to handle the new request. There was also a new event added in the App.vue for handling form fill requests by doctors.